client_credentials/pyproject.toml
[project]
name = "client-credentials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "simple-salesforce>=1.12.9",
]
